Refactor entry point (index.js) Extract Node version check and server initialisation to the startup dir Also log the number of loaded events again
