+++ startup/index.js
@@ -1,3 +1,9 @@
 import bindEvents from "./events/bindEvents";
+import initServer from "./init/initServer";
+import nodeVersionCheck from "./init/nodeVersionCheck";
 
-export default { bindEvents };
+export default {
+    bindEvents,
+    initServer,
+    nodeVersionCheck
+};
